types: fix #8068
